# repo/api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: KuberCloud ANI Core API
  version: 1.0.0
  description: |
    KuberCloud ANI Core — 基础设施平台 REST API 契约（唯一真实来源）。

    ## 文件范围与边界
    本文件（api/openapi/v1.yaml）仅包含 ANI Core 基础设施资源：
    instances（VM/Container/GPU/Sandbox/BM/K8s集群）、volumes/filesystems/objects/
    vector-stores、networks、auth/secrets/encryption、gpu-inventory/dpu-inventory、
    registry/metering/observability/notifications/audit 等。
    ANI Services 资源（模型仓库/推理/知识库/PaaS）在 api/openapi/services/v1.yaml。

    ## URL 前缀规范
    servers[0].url = "https://{host}/api/v1"（前缀在 spec 中定义，Gateway 不做路径变换）。
    版本升级时新建 api/openapi/v2.yaml，本文件路径保持稳定。

    ## SDK 生成（make gen-core-sdk）
    Go: github.com/kubercloud/ani-go | Python: pip install kubercloud-ani
    TypeScript: @kubercloud/ani-client | Java: com.kubercloud:ani-java

    ## 认证
    Bearer JWT（1小时有效）或 X-API-Key Header。
    tenant_id 从 JWT claims 提取，请求体中的 tenant_id 字段将被忽略。

    ## 错误格式
    `{ code: string, message: string, request_id: string, details?: object }`
    标准错误码：UNAUTHORIZED / FORBIDDEN / NOT_FOUND / CONFLICT / BAD_REQUEST /
    RATE_LIMIT_EXCEEDED / NOT_IMPLEMENTED / INTERNAL_ERROR

    ## 分页
    cursor 分页：`?limit=20&cursor=<token>` → `{ items: [], next_cursor: string | null }`

    ## 异步操作
    202 响应体返回 AsyncTask，Location Header 附任务 URL。
    轮询 GET /api/v1/tasks/{task_id} 或配置 Webhook。
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: ani.internal
        description: ANI Gateway 地址（支持IP地址）

...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT AccessToken（有效期1小时，通过 /auth/token 获取）
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 长期 API Key（在 Console → 设置 → API Key 中创建）

# ── 公共 Schema ───────────────────────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:        { type: string, example: "MODEL_NOT_FOUND" }
        message:     { type: string, example: "模型 qwen2.5-72b 不存在" }
        request_id:  { type: string, example: "req_01j8xabc" }
        details:     { type: object, additionalProperties: true }

    # Cursor 分页（所有列表接口统一使用此格式）
...
    CoreDevProfileInfo:
      type: object
      description: "Core dev/local profile 标记；用于区分本地联调成功与真实 provider 执行成功。"
      required: [mode, provider, real_provider]
      properties:
        mode:          { type: string, enum: [local, real] }
        provider:      { type: string }
        real_provider: { type: boolean }
        reason:        { type: string, nullable: true }
...
    StorageResourceState:
      type: string
      enum: [pending, available, failed, deleting, deleted]
    StorageVolume:
      type: object
      required: [id, tenant_id, name, size_gib, storage_class, state, created_at, updated_at]
      properties:
        id:            { type: string }
        tenant_id:     { type: string }
        name:          { type: string }
        size_gib:      { type: integer, format: int64, minimum: 1 }
        storage_class: { type: string }
        state:         { $ref: '#/components/schemas/StorageResourceState' }
        reason:        { type: string, nullable: true }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at:    { type: string, format: date-time }
        updated_at:    { type: string, format: date-time }
    StorageFilesystem:
      type: object
      required: [id, tenant_id, name, protocol, size_gib, state, created_at, updated_at]
      properties:
        id:         { type: string }
        tenant_id:  { type: string }
        name:       { type: string }
        protocol:   { type: string, enum: [nfs, cephfs] }
        size_gib:   { type: integer, format: int64, minimum: 1 }
        endpoint:   { type: string, nullable: true }
        state:      { $ref: '#/components/schemas/StorageResourceState' }
        reason:     { type: string, nullable: true }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    StorageObject:
      type: object
      required: [id, tenant_id, bucket, key, size_bytes, content_type, state, created_at, updated_at]
      properties:
        id:           { type: string }
        tenant_id:    { type: string }
        bucket:       { type: string }
        key:          { type: string }
        size_bytes:   { type: integer, format: int64, minimum: 0 }
        content_type: { type: string }
        state:        { $ref: '#/components/schemas/StorageResourceState' }
        reason:       { type: string, nullable: true }
        dev_profile:
          $ref: '#/components/schemas/CoreDevProfileInfo'
        created_at:   { type: string, format: date-time }
        updated_at:   { type: string, format: date-time }
    StorageVolumeListResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/StorageVolume' } }
        total: { type: integer }
        next_cursor: { type: string, nullable: true }
    StorageFilesystemListResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/StorageFilesystem' } }
        total: { type: integer }
        next_cursor: { type: string, nullable: true }
    StorageObjectListResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/StorageObject' } }
        total: { type: integer }
        next_cursor: { type: string, nullable: true }
    CreateStorageVolumeRequest:
      type: object
      required: [name, size_gib, idempotency_key]
      properties:
        idempotency_key: { type: string, minLength: 1, maxLength: 128, description: "客户端生成；同一 tenant_id 下 24 小时内去重" }
        name:            { type: string, minLength: 1, maxLength: 128 }
        size_gib:        { type: integer, format: int64, minimum: 1 }
        storage_class:   { type: string, default: standard }
    CreateStorageFilesystemRequest:
      type: object
      required: [name, size_gib, idempotency_key]
      properties:
        idempotency_key: { type: string, minLength: 1, maxLength: 128, description: "客户端生成；同一 tenant_id 下 24 小时内去重" }
        name:            { type: string, minLength: 1, maxLength: 128 }
        protocol:        { type: string, enum: [nfs, cephfs], default: nfs }
        size_gib:        { type: integer, format: int64, minimum: 1 }
    CreateStorageObjectRequest:
      type: object
      required: [bucket, key, idempotency_key]
      properties:
        idempotency_key: { type: string, minLength: 1, maxLength: 128, description: "客户端生成；同一 tenant_id 下 24 小时内去重" }
        bucket:          { type: string, minLength: 1, maxLength: 128 }
        key:             { type: string, minLength: 1, maxLength: 1024 }
        size_bytes:      { type: integer, format: int64, minimum: 0, default: 0 }
        content_type:    { type: string, default: application/octet-stream }
...
  responses:
    Unauthorized:
      description: 未认证或 Token 无效（code=UNAUTHORIZED）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: 无权访问该资源（code=FORBIDDEN）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: 资源不存在（code=NOT_FOUND）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: 请求参数不合法（code=BAD_REQUEST）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ── API 路径定义 ──────────────────────────────────────────────────────────────
# 说明：/demo/* 路径已迁移至 api/openapi/dev.yaml（不参与 SDK 生成）
paths:
...
  /volumes:
    get:
      operationId: listStorageVolumes
      summary: 查询块存储卷列表
      tags: [Volumes]
      x-ani-rbac-scope: "scope:volumes:read"
      parameters:
        - { name: limit,  in: query, schema: { type: integer, default: 20, minimum: 1, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: 块存储卷列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageVolumeListResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createStorageVolume
      summary: 创建块存储卷
      tags: [Volumes]
      x-ani-rbac-scope: "scope:volumes:create"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStorageVolumeRequest' }
      responses:
        "201":
          description: 块存储卷已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageVolume' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /volumes/{volume_id}:
    get:
      operationId: getStorageVolume
      summary: 查询块存储卷
      tags: [Volumes]
      x-ani-rbac-scope: "scope:volumes:read"
      parameters:
        - { name: volume_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 块存储卷
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageVolume' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteStorageVolume
      summary: 删除块存储卷
      tags: [Volumes]
      x-ani-rbac-scope: "scope:volumes:delete"
      parameters:
        - { name: volume_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 块存储卷已删除
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageVolume' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /filesystems:
    get:
      operationId: listStorageFilesystems
      summary: 查询文件存储列表
      tags: [Filesystems]
      x-ani-rbac-scope: "scope:filesystems:read"
      parameters:
        - { name: limit,  in: query, schema: { type: integer, default: 20, minimum: 1, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: 文件存储列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageFilesystemListResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createStorageFilesystem
      summary: 创建文件存储
      tags: [Filesystems]
      x-ani-rbac-scope: "scope:filesystems:create"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStorageFilesystemRequest' }
      responses:
        "201":
          description: 文件存储已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageFilesystem' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /filesystems/{filesystem_id}:
    get:
      operationId: getStorageFilesystem
      summary: 查询文件存储
      tags: [Filesystems]
      x-ani-rbac-scope: "scope:filesystems:read"
      parameters:
        - { name: filesystem_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 文件存储
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageFilesystem' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteStorageFilesystem
      summary: 删除文件存储
      tags: [Filesystems]
      x-ani-rbac-scope: "scope:filesystems:delete"
      parameters:
        - { name: filesystem_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 文件存储已删除
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageFilesystem' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
  /objects:
    get:
      operationId: listStorageObjects
      summary: 查询对象元数据列表
      tags: [ObjectStore]
      x-ani-rbac-scope: "scope:objects:read"
      parameters:
        - { name: limit,  in: query, schema: { type: integer, default: 20, minimum: 1, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: 对象元数据列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageObjectListResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createStorageObject
      summary: 创建对象元数据
      tags: [ObjectStore]
      x-ani-rbac-scope: "scope:objects:create"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateStorageObjectRequest' }
      responses:
        "201":
          description: 对象元数据已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageObject' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
  /objects/{object_id}:
    get:
      operationId: getStorageObject
      summary: 查询对象元数据
      tags: [ObjectStore]
      x-ani-rbac-scope: "scope:objects:read"
      parameters:
        - { name: object_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 对象元数据
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageObject' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteStorageObject
      summary: 删除对象元数据
      tags: [ObjectStore]
      x-ani-rbac-scope: "scope:objects:delete"
      parameters:
        - { name: object_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: 对象元数据已删除
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StorageObject' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── 向量存储（Core API Alpha）──────────────────────────────────────────────
...